5/13
Now registering collisions, which renders a blast image when colliding, Reduced rock image rendering due to issue where you wouldn't be able to register collisions, Issues with an attempted system to be knocked backwards when colliding but very buggy so not implemented as of now. Added health variable and will start implementing.